+++ src/test/java/net/bzethmayr/gigantspinosaurus/usage/pipelines/VideoMarringPipelineTest.java
@@ -2,6 +2,8 @@ package net.bzethmayr.gigantspinosaurus.usage.pipelines;
 
 import net.bzethmayr.gigantspinosaurus.model.mar.ExposesMar;
 import net.bzethmayr.gigantspinosaurus.usage.*;
+import net.bzethmayr.gigantspinosaurus.usage.defaults.DefaultEnvironments;
+import net.bzethmayr.gigantspinosaurus.usage.defaults.windows.WindowsEnvironments;
 import net.bzethmayr.gigantspinosaurus.usage.images.CrossFormatDecoder;
 import net.bzethmayr.gigantspinosaurus.usage.images.CrossFormatDecoder.Raster;
 import net.bzethmayr.gigantspinosaurus.usage.images.TestsWithImages;
@@ -19,7 +21,9 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
 import static net.bzethmayr.gigantspinosaurus.usage.BindsConstructors.defaultConstructors;
-import static net.bzethmayr.gigantspinosaurus.usage.BindsEnvironment.desktopEnvironment;
+import static net.bzethmayr.gigantspinosaurus.usage.defaults.DefaultEnvironments.desktopEnvironment;
+import static net.bzethmayr.gigantspinosaurus.usage.defaults.windows.WindowsEnvironments.windowsPermanentEnvironment;
+import static net.bzethmayr.gigantspinosaurus.usage.video.VideoMarringCoordinator.blockingCoordinator;
 import static org.junit.jupiter.api.Assertions.*;
 
 class VideoMarringPipelineTest implements TestsWithImages {
@@ -36,10 +40,9 @@ class VideoMarringPipelineTest implements TestsWithImages {
 
         final var preparation = new QrMarkEmbedder(raster.width(), raster.height());
         final var pipeline = new BindsMarkingPipeline(reduction, preparation, preparation);
-        final var underTest = new VideoMarring(
-                BindsConstructors.defaultConstructors(),
-                BindsEnvironment.desktopEnvironment(),
-                pipeline, VideoMarringCoordinator.blockingCoordinator(), 1, 1);
+        final var underTest = new VideoMarring(defaultConstructors(),
+                windowsPermanentEnvironment(),
+                pipeline, blockingCoordinator(), 1, 1);
 
         final ByteBuffer frame = raster.toBuffer();
         final var media = underTest.mediaFrame();
@@ -93,10 +96,8 @@ class VideoMarringPipelineTest implements TestsWithImages {
 
         final var embedder = new QrMarkEmbedder(w, h);
         final var pipeline = new BindsMarkingPipeline(reduction, embedder, embedder);
-        final var underTest = new VideoMarring(
-                BindsConstructors.defaultConstructors(),
-                BindsEnvironment.desktopEnvironment(),
-                pipeline, VideoMarringCoordinator.blockingCoordinator(), 1, 1);
+        final var underTest = new VideoMarring(defaultConstructors(), desktopEnvironment(),
+                pipeline, blockingCoordinator(), 1, 1);
 
         final var media = underTest.mediaFrame();
         final var background = underTest.background();
@@ -139,10 +140,8 @@ class VideoMarringPipelineTest implements TestsWithImages {
 
             final var embedder = new QrMarkEmbedder(raster.width(), raster.height());
             final var pipeline = new BindsMarkingPipeline(reduction, embedder, embedder);
-            final var underTest = new VideoMarring(
-                    BindsConstructors.defaultConstructors(),
-                    BindsEnvironment.desktopEnvironment(),
-                    pipeline, VideoMarringCoordinator.blockingCoordinator(), 1, 1);
+            final var underTest = new VideoMarring(defaultConstructors(), desktopEnvironment(),
+                    pipeline, blockingCoordinator(), 1, 1);
 
             final ByteBuffer frame = raster.toBuffer();
             final var media = underTest.mediaFrame();
@@ -186,10 +185,8 @@ class VideoMarringPipelineTest implements TestsWithImages {
 
             final var embedder = new QrMarkEmbedder(w, h);
             final var pipeline = new BindsMarkingPipeline(reduction, embedder, embedder);
-            final var underTest = new VideoMarring(
-                    BindsConstructors.defaultConstructors(),
-                    BindsEnvironment.desktopEnvironment(),
-                    pipeline, VideoMarringCoordinator.blockingCoordinator(), 1, 1);
+            final var underTest = new VideoMarring(defaultConstructors(), desktopEnvironment(),
+                    pipeline, blockingCoordinator(), 1, 1);
 
             final var media = underTest.mediaFrame();
             final var background = underTest.background();
@@ -225,7 +222,7 @@ class VideoMarringPipelineTest implements TestsWithImages {
         final Raster raster = CrossFormatDecoder.decode(imagePath);
 
         final var ctors = defaultConstructors();
-        final var env = desktopEnvironment();
+        final var env = windowsPermanentEnvironment();
         try (var gpu = new VulkanRoot()) {
             final var reduction = new VideoReduction(gpu, raster.width(), raster.height());
             final var creation = new MarCreation(ctors, env);